// frontend/src/lib/api/types.gen.ts
export type Health = {
    status: 'ok';
    version: string;
    kraftVersion: string;
};

export type Error = {
    code: string;
    message: string;
};

export type Machine = {
    id?: string;
    name: string;
    kernel?: string;
    args?: string;
    created?: string;
    status: string;
    memory?: string;
    ports?: string;
    ip?: string;
    pid?: string;
    platform: string;
    architecture: string;
};

export type Package = {
    name: string;
    version: string;
    type: string;
    format?: string;
    platform?: string;
    architecture?: string;
    size?: string;
    created?: string;
    updated?: string;
    pulled?: string;
    description?: string;
    origin?: string;
    index?: string;
    manifest?: string;
};

export type Network = {
    machineId?: string;
    name: string;
    network?: string;
    driver: string;
    status: string;
};

export type Volume = {
    id?: string;
    name: string;
    driver: string;
    status: string;
    source?: string;
    destination?: string;
};